requirements.txt
uagents==0.4.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
google-generativeai>=0.3.0
requests
flask 